Major services file structure update
