--- src/services/translatePageType.js
@@ -1,14 +1,0 @@
-export const translatePageType = (pageType) => {
-    switch (pageType){
-        case 'post':
-            return 'blog';
-        case 'course':
-            return 'courses';
-        case 'podcast':
-            return 'podcasts';
-        case 'product':
-            return 'store';
-        default:
-            return null;
-    }
-};
